Add separate popup component
