+++ packages/website/src/routes/Homepage/Map/index.tsx
@@ -1,24 +1,12 @@
 import React, { useState, useRef, useEffect } from "react";
 import { useSelector } from "react-redux";
-import { Map, TileLayer, Marker, Popup } from "react-leaflet";
+import { Map, TileLayer, Marker } from "react-leaflet";
 import L from "leaflet";
-import {
-  createStyles,
-  withStyles,
-  WithStyles,
-  Theme,
-  Typography,
-  Card,
-  CardHeader,
-  CardContent,
-  Grid,
-  Button,
-} from "@material-ui/core";
-import { Link } from "react-router-dom";
+import { createStyles, withStyles, WithStyles } from "@material-ui/core";
 
 import { reefsListSelector } from "../../../store/Reefs/reefsListSlice";
 import { Reef } from "../../../store/Reefs/types";
-import { colors } from "../../../layout/App/theme";
+import Popup from "./popup";
 
 const marker = require("../../../assets/buoy.png");
 
@@ -69,65 +57,7 @@ const HomepageMap = ({ classes }: HomepageMapProps) => {
                   reef.polygon.coordinates[0],
                 ]}
               >
-                <Popup closeButton={false} className={classes.popup}>
-                  <Card>
-                    <CardHeader
-                      className={classes.popupHeader}
-                      title={reef.name}
-                      subheader={reef.region}
-                    />
-                    <CardContent>
-                      <Grid container item xs={12}>
-                        <Grid container justify="flex-start" item xs={6}>
-                          <Typography variant="caption" color="textSecondary">
-                            TEMP AT 25M
-                          </Typography>
-                          <Typography
-                            style={{ color: colors.lightBlue }}
-                            variant="h5"
-                            color="textSecondary"
-                          >
-                            31.8 &#8451;
-                          </Typography>
-                        </Grid>
-                        <Grid container justify="flex-end" item xs={6}>
-                          <Typography variant="caption" color="textSecondary">
-                            DEG. HEAT. DAYS
-                          </Typography>
-                          <Typography
-                            style={{ color: "purple" }}
-                            variant="h5"
-                            color="textSecondary"
-                          >
-                            58
-                          </Typography>
-                        </Grid>
-                      </Grid>
-                      <Grid
-                        style={{ margin: "1rem 0 1rem 0" }}
-                        container
-                        justify="flex-start"
-                        item
-                        xs={12}
-                      >
-                        <Grid item>
-                          <Link
-                            style={{ color: "inherit", textDecoration: "none" }}
-                            to={`/reefs/${reef.id}`}
-                          >
-                            <Button
-                              size="small"
-                              variant="outlined"
-                              color="primary"
-                            >
-                              EXPLORE
-                            </Button>
-                          </Link>
-                        </Grid>
-                      </Grid>
-                    </CardContent>
-                  </Card>
-                </Popup>
+                <Popup reef={reef} />
               </Marker>
             );
           }
@@ -137,17 +67,11 @@ const HomepageMap = ({ classes }: HomepageMapProps) => {
   );
 };
 
-const styles = (theme: Theme) =>
+const styles = () =>
   createStyles({
     map: {
       height: "100%",
       width: "100%",
-    },
-    popup: {
-      width: "12vw",
-    },
-    popupHeader: {
-      backgroundColor: theme.palette.primary.main,
     },
   });
